backend/requirements.txt
tensorflow==2.15.0
numpy
flask
flask-cors
pillow